More sensible integration, less warnings
Tidied up some things to remove compiler warnings.
AO integration is now over the hemisphere, and answer is properly
normalised.
SH integration now no longer computes SH func if value to be integrated
is 0 (should be more efficient).

+++ src/SHMat.cpp
@@ -29,7 +29,7 @@ std::vector<float> SHMat::operator * (const std::vector<float>& p)
 
 	std::vector<float> ans;
 
-	for(int i = 0; i < blocks.size(); ++i)
+	for(size_t i = 0; i < blocks.size(); ++i)
 	{
 		std::vector<float> subVec(p.begin() + (i*i), p.begin() + (i+1)*(i+1));
 		std::vector<float> subProd = blocks[i] * subVec;
@@ -48,7 +48,7 @@ std::vector<glm::vec4> SHMat::operator * (const std::vector<glm::vec4>& p)
 
 	std::vector<glm::vec4> ans;
 
-	for(int i = 0; i < blocks.size(); ++i)
+	for(size_t i = 0; i < blocks.size(); ++i)
 	{
 		std::vector<glm::vec4> subVec(p.begin() + (i*i), p.begin() + (i+1)*(i+1));
 		std::vector<glm::vec4> subProd = blocks[i] * subVec;
